WIP: start shifting version setting logic into requests
This is enough to get the examples/consumergroup example to work.

Move the logic for setting a request's version from a KafkaVersion
struct into the source file for each request. The motivation for doing
this is that special cases need to be added for the new `Automatic`
version (which can't be compared against other versions), so it helps to
consolidate all the logic that cares about the KafkaVersion into a few
places.

These changes assume that it's reasonable to populate all possible
fields of a request struct, then let the encoding/decoding process
decide which fields need to be transmitted based on the request's
version. So far this has worked out, but could prove problematic if
calculating a field is expensive, or might rely on some piece of data
that isn't always available.