WIP: Migrate async_producer.go
As per the comment, there is one place where setting the version to
Automatic and connecting to a 0.10 broker will introduce a small (10
byte) inaccuracy in the client-side detection of breaching the
Producer.MaxMessageSize limit. This seems like an acceptable trade-off
in a niche case, compared to the re-work require to push this check down
into code that has the context of an established broker connection.